+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- strip manageWorkspaces from consumer bit.json by default
+- bug fix - when removing a missing component from a remote, it was showing an obscure error "id.toStringWithoutVersion is not a function"
+
 ## [0.12.0-dev.29] - 2018-01-16
 
 - change commit hash to tag version in bit logs